backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
joblib
httpx
PyJWT
python-dotenv
chromadb
